requirements.txt
nextcord
python-dotenv
mysql-connector-python
captcha